Data  Set (Train)
This data set is downloaded directly for the Data Brick using SQL.
It contains about 1,048,576 rows of data and due to its sheer size, I have uploaded a zip copy.

Don't worry we don't use this huge data direct. It goes through a lot of data processing and cleaning methods. 

Unfortunately, even after zipping the file I am unable to upload it to GitHub. So finally attached an image. 

haha, sorry for the inconvenience guys!