Test Data Set
And lastly here is the is an image of the test data because as you can understand this data is large as well.
Now you might ask, why not proceed with a small data?
Ans: These predictions are done voyage vise, which means that vessel will be at sea for an appreciable amount of time. 
Hence when we do Power prediction we do it for the entire voyage.